Move FileLoaderSection to opengever.bundle
