+++ opengever/bundle/cfgs/oggbundle.cfg
@@ -39,7 +39,7 @@ blueprint = opengever.setup.resolveguid
 blueprint = opengever.setup.constructor
 
 [fileloader]
-blueprint = opengever.setup.fileloader
+blueprint = opengever.bundle.fileloader
 
 [map-local-roles]
 blueprint = opengever.bundle.map_local_roles
