+++ opengever/bundle/tests/test_section_fileloader.py
@@ -4,8 +4,8 @@ from ftw.builder import Builder
 from ftw.builder import create
 from opengever.bundle.sections.bundlesource import BUNDLE_PATH_KEY
 from opengever.bundle.sections.bundlesource import JSON_STATS_KEY
-from opengever.setup.sections.fileloader import FileLoaderSection
-from opengever.setup.tests import MockTransmogrifier
+from opengever.bundle.sections.fileloader import FileLoaderSection
+from opengever.bundle.tests import MockTransmogrifier
 from opengever.testing import FunctionalTestCase
 from pkg_resources import resource_filename
 from plone import api
