+++ opengever/setup/tests/test_section_resolveguid.py
@@ -1,10 +1,10 @@
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
+from opengever.bundle.tests import MockTransmogrifier
 from opengever.setup.sections.resolveguid import DuplicateGuid
 from opengever.setup.sections.resolveguid import MissingGuid
 from opengever.setup.sections.resolveguid import MissingParent
 from opengever.setup.sections.resolveguid import ResolveGUIDSection
-from opengever.setup.tests import MockTransmogrifier
 from opengever.testing import FunctionalTestCase
 from zope.interface.verify import verifyClass
 from zope.interface.verify import verifyObject
