+++ opengever/setup/transmogrifier.zcml
@@ -85,9 +85,4 @@
     name="opengever.setup.constructor"
     />
 
-  <utility
-    component=".sections.fileloader.FileLoaderSection"
-    name="opengever.setup.fileloader"
-    />
-
 </configure>
